requirements.txt
fastapi[all]
prometheus-client
environ-config
